rolesAuth in Dashboards

+++ app/services/DashboardService.java
@@ -24,6 +24,7 @@ import java.util.stream.Collectors;
 
 import static play.mvc.Controller.ctx;
 import static utils.Constants.READ_ACL;
+import static utils.Constants.WRITE_ACL;
 
 public class DashboardService {
     private MongoDatabase database;
@@ -37,48 +38,80 @@ public class DashboardService {
         return CompletableFuture.supplyAsync(() -> {
             MongoCollection<Dashboard> dashboards = database.getCollection(collectionName, Dashboard.class);
             FindIterable<Dashboard> find = dashboards.find();
-
-            List<Integer> access = new ArrayList<>();
-            List<Integer> roles = user.getRoles().stream().map(next -> next.getId()).collect(Collectors.toList());
+            List<String> access = new ArrayList<>();
+            List<String> roles = user.getRoles().stream().map(next -> next.getId()).collect(Collectors.toList());
             System.out.println(roles);
-//            access.addAll(roles);
-//            access.add(user.getId());
-//            access.add("*");
+            access.addAll(roles);
+            access.add(user.getId().toString());
+            access.add("*");
             // add filters, based on user access
             ArrayList<Dashboard> dash = find
                     .filter(Filters.or(Filters.in(READ_ACL, access), Filters.size(READ_ACL, 0)))
                     .into(new ArrayList<>());
-
             if (dash == null)
                 throw new CompletionException(new RequestException(Http.Status.NOT_FOUND, "Nothing founded"));
             return dash;
         }, context);
     }
 
-    public CompletableFuture<Dashboard> findById(String t, Executor context) {
+    public CompletableFuture<Dashboard> findById(String t, Executor context,User user) {
         return CompletableFuture.supplyAsync(() -> {
             MongoCollection<Dashboard> dashboards = database.getCollection(collectionName, Dashboard.class);
             BasicDBObject query = new BasicDBObject();
             query.put("_id", new ObjectId(t));
-            Dashboard dash = dashboards.find(query).first();
+
+            List<String> access = new ArrayList<>();
+            List<String> roles = user.getRoles().stream().map(next -> next.getId()).collect(Collectors.toList());
+            System.out.println(roles);
+            access.addAll(roles);
+            access.add(user.getId().toString());
+            access.add("*");
+
+            Dashboard dash = dashboards.find(query).filter(Filters.or(Filters.in(READ_ACL, access), Filters.size(READ_ACL, 0))).first();
             if (dash == null)
                 throw new CompletionException(new RequestException(Http.Status.NOT_FOUND, "Nothing founded"));
             return dash;
         }, context);
     }
 
-    public CompletableFuture<Dashboard> save(Dashboard which, Executor context) {
+    public CompletableFuture<Dashboard> save(Dashboard which, Executor context,User user) {
         return CompletableFuture.supplyAsync(() -> {
             MongoCollection<Dashboard> dashboards = database.getCollection(collectionName, Dashboard.class);
+            //todo check up how to know if this user has access to create new dashboard :) :P :/
+            List<String> access = new ArrayList<>();
+            List<String> roles = user.getRoles().stream().map(next -> next.getId()).collect(Collectors.toList());
+            System.out.println(roles);
+            access.addAll(roles);
+            access.add(user.getId().toString());
+            access.add("*");
+
+            Dashboard dash = dashboards.find().filter(Filters.or(Filters.in(WRITE_ACL, access), Filters.size(WRITE_ACL, 0))).first();
+
             which.setId(new ObjectId());
             dashboards.insertOne(which);
             return which;
         }, context);
     }
 
-    public CompletableFuture<Dashboard> update(Dashboard which, Executor context) {
+    public CompletableFuture<Dashboard> update(Dashboard which, Executor context,User user) {
         return CompletableFuture.supplyAsync(() -> {
             MongoCollection<Dashboard> dashboards = database.getCollection(collectionName, Dashboard.class);
+
+            List<String> access = new ArrayList<>();
+            List<String> roles = user.getRoles().stream().map(next -> next.getId()).collect(Collectors.toList());
+            System.out.println(roles);
+            access.addAll(roles);
+            access.add(user.getId().toString());
+            access.add("*");
+
+            BasicDBObject query = new BasicDBObject();
+            query.put("_id", new ObjectId(which.getId().toString()));
+
+            Dashboard dash = dashboards.find(query).filter(Filters.or(Filters.in(WRITE_ACL, access), Filters.size(WRITE_ACL, 0))).first();
+            if(dash == null){
+                throw new CompletionException(new RequestException(Http.Status.UNAUTHORIZED,"Permission denied"));
+            }
+
             UpdateResult updateResult = dashboards.updateOne(Filters.eq("_id", which.getId()), new BasicDBObject("$set", which));
             if (updateResult.getModifiedCount() > 0)
                 return which;
@@ -86,11 +119,25 @@ public class DashboardService {
         }, context);
     }
 
-    public CompletableFuture<DeleteResult> delete(String id, Executor context) {
+    public CompletableFuture<DeleteResult> delete(String id, Executor context,User user) {
         return CompletableFuture.supplyAsync(() -> {
             MongoCollection<Dashboard> dashboards = database.getCollection(collectionName, Dashboard.class);
+
+            List<String> access = new ArrayList<>();
+            List<String> roles = user.getRoles().stream().map(next -> next.getId()).collect(Collectors.toList());
+            System.out.println(roles);
+            access.addAll(roles);
+            access.add(user.getId().toString());
+            access.add("*");
+
             BasicDBObject query = new BasicDBObject();
             query.put("_id", new ObjectId(id));
+
+            Dashboard dash = dashboards.find(query).filter(Filters.or(Filters.in(WRITE_ACL, access), Filters.size(WRITE_ACL, 0))).first();
+            if(dash == null){
+                throw new CompletionException(new RequestException(Http.Status.UNAUTHORIZED,"Permission denied"));
+            }
+
             DeleteResult deleteResult = dashboards.deleteOne(query);
             if (deleteResult.getDeletedCount() > 0)
                 return deleteResult;
